+++ src/widgets/alignment.zig
@@ -1,6 +1,6 @@
 const Window = @import("../Window.zig");
 
-pub fn center(parent: Window, cols: usize, rows: usize) Window {
+pub fn center(parent: Window, cols: u16, rows: u16) Window {
     const y_off = (parent.height / 2) -| (rows / 2);
     const x_off = (parent.width / 2) -| (cols / 2);
     return parent.child(.{
